# server/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "activities-server",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
		"/activity": {
			"get": {
				"operationId": "GetActivity",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ActivityDto"
								}
							}
						}
					}
				},
				"description": "Gets new random activity, based on current user preferences if a user was set",
				"tags": [
					"Activity"
				],
				"security": [],
				"parameters": []
			}
		},
		"/user": {
			"post": {
				"operationId": "PostUser",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Sets a new current user",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"components": {
		"schemas": {
			"Accessibility": {
				"enum": [
					"Low",
					"Medium",
					"High"
				],
				"type": "string"
			},
			"Price": {
				"enum": [
					"Free",
					"Low",
					"High"
				],
				"type": "string"
			},
			"ActivityDto": {
				"properties": {
					"activity": {
						"type": "string"
					},
					"accessibility": {
						"$ref": "#/components/schemas/Accessibility"
					},
					"type": {
						"type": "string"
					},
					"participants": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"$ref": "#/components/schemas/Price"
					},
					"link": {
						"type": "string"
					},
					"key": {
						"type": "string"
					}
				},
				"required": [
					"activity",
					"accessibility",
					"type",
					"participants",
					"price",
					"link",
					"key"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"name": {
						"type": "string"
					},
					"accessibility": {
						"$ref": "#/components/schemas/Accessibility"
					},
					"price": {
						"$ref": "#/components/schemas/Price"
					}
				},
				"required": [
					"name",
					"accessibility",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	}
}